backtest and realtest still have something to change about orderFunction

+++ AnyQuant_Web/src/main/java/function/order/ShareVO.java
@@ -3,9 +3,9 @@ package function.order;
 public class ShareVO {
 	public int order;
 	public String siid;
-	public int share;
+	public double share;
 	public double price;
-	public ShareVO(int order, String siid, int share,double price) {
+	public ShareVO(int order, String siid, double share,double price) {
 		super();
 		this.order = order;
 		this.siid = siid;
@@ -24,10 +24,10 @@ public class ShareVO {
 	public void setSiid(String siid) {
 		this.siid = siid;
 	}
-	public int getShare() {
+	public double getShare() {
 		return share;
 	}
-	public void setShare(int share) {
+	public void setShare(double share) {
 		this.share = share;
 	}
 	public double getPrice() {
